requirements.txt
annotated-types==0.7.0
anyio==4.4.0
backoff==1.11.1
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
dataclasses-json==0.6.6
degiro-connector==3.0.0
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.1
fake-useragent==1.5.1
google-cloud==0.34.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
install==1.3.5
marshmallow==3.21.2
mypy-extensions==1.0.0
onetimepass==1.0.1
openai==1.30.3
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
polars==0.20.27
protobuf==5.26.1
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
ratelimiter==1.2.0
requests==2.32.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tomli==2.0.1
tqdm==4.62.3
typeguard==4.2.1
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==1.26.18
wrapt==1.16.0
zipp==3.18.2
